+++ sdk/python/pulumi_gcp/runtimeconfig/__init__.py
@@ -4,4 +4,4 @@
 
 # Export this package's modules as members:
 from .config import *
-from .variavble import *
+from .variable import *